Struct iata_bcbp::Bcbp

pub struct Bcbp { /* fields omitted */ }

Methods

impl Bcbp

pub fn legs(&self) -> &[Leg]

All legs encoded into the boarding pass. At least one needs to be present to form a valid boarding pass.

pub fn security_data(&self) -> &SecurityData

A reference to the optional security data used to verify a boarding pass was not tamptered with.

pub fn electronic_ticket_indicator(&self) -> char

Used to differentiate between an electronic ticket ('E') and another type of travel document. Values are defined in Resolution 792. A space indicates the field is not set.

pub fn passenger_description(&self) -> Option<char>

This describes the passenger. Values are defined in Resolution 792. Spaces indicate the field is not set.

pub fn passenger_name(&self) -> &str

The name of the passenger. Up to 20 characters, left-aligned, space padded. The format is LAST_NAME/FIRST_NAME[TITLE]. There is no separator between the first name and the title, and no indication a title is present. Certain names have characters which cannot be translated and special handling may be required. Spaces indicate the field is not set.

pub fn source_of_check_in(&self) -> Option<char>

This field reflects channel in which the customer initiated check-in. Values are defined in Resolution 792 Attachment C. Spaces indicate the field is not set.

pub fn source_of_boarding_pass_issuance(&self) -> Option<char>

This field reflects channel which issued the boarding pass. Values are defined in Resolution 792. Spaces indicate the field is not set.

pub fn date_of_issue_of_boarding_pass(&self) -> Option<&str>

Optionally the 4-digit Julian date representing when the boarding pass was issued. The first digit is the last digit of the year and the next three represent the number of days elapsed. For example: "6001" represnts January 1, 2016. "6366" represaents December 31, 2016 (a leap year). Spaces indicate the field is not set.

pub fn document_type(&self) -> Option<char>

The type of the document, 'B' indicating a boarding pass. Spaces indicate the field is not set.

pub fn airline_designator_of_boarding_pass_issuer(&self) -> Option<&str>

Airline code of the boarding pass issuer. Two-character and three-letter IATA carrier designators are permitted and the string is left-justified and space padded. Spaces indicate the field is not set.

pub fn baggage_tag_license_plate_numbers(&self) -> Option<&str>

This field allows carriers to populate baggage tag numbers and the number of consecutive bags. This 13-character fiels is divided into: 0: '0' for interline tag, '1' for fall-back tag, '2' for interline rush tag. 2... 4: carrier numeric code. 5...10: carrier initial tag number with leading zeroes. 11...13: number of consecutive bags (up to 999). Spaces indicate the field is not set.

pub fn first_non_consecutive_baggage_tag_license_plate_numbers(
&self
) -> Option<&str>

This field allows carriers who handle non-sequential bags to include a second set of them in the boarding pass data in in the same format as baggage_tag_license_plate_numbers. Spaces indicate the field is not set.

pub fn second_non_consecutive_baggage_tag_license_plate_numbers(
&self
) -> Option<&str>

This field allows carriers who handle non-sequential bags to include a third set of them in the boarding pass data in in the same format as baggage_tag_license_plate_numbers. Spaces indicate the field is not set.