Struct iata_bcbp::Leg

pub struct Leg { /* fields omitted */ }

Methods

impl Leg

pub fn operating_carrier_pnr_code(&self) -> &str

An alphanumeric string of up to 6 characters, left-aligned, space-padded. This is the Passenger Name Record used to identify the booking in the reservation system of the operating carrier.

pub fn from_city_airport_code(&self) -> &str

Three-letter or four-letter IATA code of the origin airport. Spaces indicate the field is not set. Any other values are invalid.

pub fn to_city_airport_code(&self) -> &str

Three-letter or four-letter IATA code of the destination airport. Spaces indicate the field is not set. Any other values are invalid.

pub fn marketing_carrier_designator(&self) -> Option<&str>

Airline code of the marketing carrier, which can be the same as the operating carrier. Two-character and three-letter IATA carrier designators are permitted and the string is left-justified and space padded. Spaces indicate the field is not set. Any other values are invalid.

pub fn frequent_flyer_airline_designator(&self) -> Option<&str>

Airline code associated with the frequent flyer number. Two-character and three-letter IATA carrier designators are permitted and the string is left-justified and space padded. Spaces indicate the field is not set. Any other values are invalid.

pub fn frequent_flyer_number(&self) -> Option<&str>

2 character or 3 letter airline designator followed by up to 13 numerics or alphanumerics, or 16 numerics if the FFN is 16 digits. Spaces indicate the field is not set. Any other values are invalid.

pub fn id_ad_indicator(&self) -> Option<char>

Values are defined in Resolution 792. Spaces indicate the field is not set. Any other values are invalid.

pub fn operating_carrier_designator(&self) -> &str

Airline code of the operating carrier, which can be the same as the marketing carrier. Two-character and three-letter IATA carrier designators are permitted and the string is left-justified and space padded. Spaces indicate the field is not set. Any other values are invalid.

pub fn flight_number(&self) -> &str

A flight number comprised of four numeric characters followed by an optional alphabetic suffix. This refers to the operating carrier. Spaces indicate the field is not set.

pub fn date_of_flight(&self) -> &str

The Julian date code for the flight. The 3-digit number reflects the day of the year beginning with '0'. The year is to be inferred. Spaces indicate the field is not set.

pub fn compartment_code(&self) -> char

IATA compartment code indiciating the class of service. Values are defined in Resolution 792. A space indicates the field is not set. Any other values are invalid.

pub fn seat_number(&self) -> &str

Seat number of the passenger. Usually 3 numerics followed by a single alphabetic. In the case of infants, can be any 4 ASCII characters, often 'INF '. Spaces indicate the field is not set.

pub fn check_in_sequence_number(&self) -> &str

Check-in sequence number. Usually 4 numerics followed by an optional alpha or blank, however in the case of infants, the format is defined by the host system and can be any 5 ASCII characters.

pub fn passenger_status(&self) -> char

The status of the passenger. Field values are defined in Resolution 792. A space indicates the field is not set.

pub fn airline_numeric_code(&self) -> Option<&str>

The three-digit airline numeric code. This is also the first three digits of the eTicket number. Spaces indicate the field is not set.

pub fn document_form_serial_number(&self) -> Option<&str>

The ten-digit DSN. This is also the last ten digits of the eTicket number. Spaces indicate the field is not set.

pub fn selectee_indicator(&self) -> Option<char>

This field is used by certain agencies to demarcate individuals requiring extra screening. Although a conditional field, it is now required as of Resolotion 792 Version 6 when travel involves the United States. Values '0', '1', or '3' determine the type of screening the passenger will receive at US airports. A space indicates the field is not set.

pub fn international_document_verification(&self) -> Option<char>

This field is used by carriers to identify passengers requiring document verification. Connected to the display of the 'DOCS OK' string on international boarding passes.

pub fn fast_track(&self) -> Option<char>

Indicates if the passenger is eligible for fast track. If 'Y', the passenger is eligible, 'N' if not, ' ' if not set. Any other values are invalid.

pub fn free_baggage_allowance(&self) -> Option<&str>

Three characters, unstructured, left-aligned and space padded, indicating how much baggage passengers are able to take with them free of charge. Spaces indicate the field is not set.

pub fn airline_individual_use(&self) -> Option<&str>

Optional unstructured data for airline individual use. Content frequently includes frequent flyer tier, passenger preferences, etc.